Experimented with a replacement for 'yield with timer' practice
Godot's tutorials recommend using 'yield(get_tree().create_timer())' but this
causes errors. I found a replacement suggested in the godot forums
(https://godotengine.org/qa/8656/how-properly-stop-yield-from-resuming-after-the-class-freed)
but the replacement still causes 'ObjectDB instances leaked' warnings when
closing the problem.

It seems there is currently no good way in GDScript to wait for 3
seconds, without making use of signals and timers, and breaking a 5-line
"do this, wait, do this, wait, do this" method into 3 different methods.
They are looking to introduce one in Godot 4.0, but until then I think
it's best to leave this as is.